almost there

+++ CashmaticApp/Windows/MainWindow.xaml.cs
@@ -34,10 +34,13 @@ namespace CashmaticApp.Windows
             string pathTodirectoryBills = baseDir + "Bills";
             string pathTodirectoryDebugLog = baseDir + "DebugLog";
             string pathTodirectoryReceipts = baseDir + "Receipts";
+            string pathTodirectoryBalances = baseDir + "Balances";
+
             Debug.Log("CashmaticApp", "Basedir: "+baseDir);
             Debug.Log("CashmaticApp", "pathTodirectoryBills: " + pathTodirectoryBills);
             Debug.Log("CashmaticApp", "pathTodirectoryDebugLog: " + pathTodirectoryDebugLog);
             Debug.Log("CashmaticApp", "pathTodirectoryReceipts: " + pathTodirectoryReceipts);
+            Debug.Log("CashmaticApp", "pathTodirectoryBalances: " + pathTodirectoryBalances);
 
 
             try
@@ -45,6 +48,7 @@ namespace CashmaticApp.Windows
                 System.IO.Directory.CreateDirectory(pathTodirectoryBills);
                 System.IO.Directory.CreateDirectory(pathTodirectoryDebugLog);
                 System.IO.Directory.CreateDirectory(pathTodirectoryReceipts);
+                System.IO.Directory.CreateDirectory(pathTodirectoryBalances);
 
                 Global.ready2orderAuthorization = ConfigurationManager.AppSettings["ready2orderAuthorization"];
                 Global.pandaParkenAuthorization = ConfigurationManager.AppSettings["pandaParkenAuthorization"];
@@ -63,19 +67,53 @@ namespace CashmaticApp.Windows
                 Global.thankYouTimer = double.Parse(ConfigurationManager.AppSettings["thankYouTimer"]);
                 Global.sixPrintReceiptWidth = int.Parse(ConfigurationManager.AppSettings["sixPrintReceiptWidth"]);
 
-                if (Global.isCashPayment) { 
-                    File.Delete(Global.cashmaticBasePath+"connected");
+                string balanceTimeFrom = ConfigurationManager.AppSettings["balanceTimeFrom"];
+                string balanceTimeTo = ConfigurationManager.AppSettings["balanceTimeTo"];
 
-                    System.Diagnostics.Process[] pname = System.Diagnostics.Process.GetProcessesByName("Cashmatic");
-                    if (pname.Length == 0) { 
-                        System.Diagnostics.ProcessStartInfo PSI = new System.Diagnostics.ProcessStartInfo();
-                        PSI.FileName = System.IO.Path.GetFileName(ConfigurationManager.AppSettings["CashmaticAppPath"]);
-                        PSI.WorkingDirectory = System.IO.Path.GetDirectoryName(ConfigurationManager.AppSettings["CashmaticAppPath"]);
-                        PSI.WindowStyle = System.Diagnostics.ProcessWindowStyle.Hidden;
-                        PSI.WindowStyle = System.Diagnostics.ProcessWindowStyle.Minimized;
-                        System.Diagnostics.Process.Start(PSI);
+                //if number, must be between 1 and 24
+                bool isNumeric = int.TryParse(balanceTimeFrom, out int timeFrom);
+                if(isNumeric && timeFrom >=1 && timeFrom<=24)
+                {
+                    Global.BalanceTimeFrom = timeFrom;
+                }
+                //if number, must be between 1 and 24
+                isNumeric = int.TryParse(balanceTimeTo, out int timeTo);
+                if (isNumeric && timeTo >= 1 && timeTo <= 24)
+                {
+                    Global.BalanceTimeTo= timeTo;
+                }
+                //if time to is smaller than time from, reverse the time order (from becomes to and vice versa)
+                if(timeTo<=timeFrom)
+                {
+                    Global.BalanceTimeFrom = timeTo;
+                    Global.BalanceTimeTo = timeFrom;
+                }
+
+                Global.BalanceTimer = double.Parse( ConfigurationManager.AppSettings["balanceTimer"]);
+
+                if(File.Exists(ConfigurationManager.AppSettings["CashmaticAppPath"]))
+                {
+                    if (Global.isCashPayment)
+                    {
+                        File.Delete(Global.cashmaticBasePath + "connected");
+                        System.Diagnostics.Process[] pname = System.Diagnostics.Process.GetProcessesByName("Cashmatic");
+                        if (pname.Length == 0)
+                        {
+                            System.Diagnostics.ProcessStartInfo PSI = new System.Diagnostics.ProcessStartInfo();
+                            PSI.FileName = System.IO.Path.GetFileName(ConfigurationManager.AppSettings["CashmaticAppPath"]);
+                            PSI.WorkingDirectory = System.IO.Path.GetDirectoryName(ConfigurationManager.AppSettings["CashmaticAppPath"]);
+                            PSI.WindowStyle = System.Diagnostics.ProcessWindowStyle.Hidden;
+                            PSI.WindowStyle = System.Diagnostics.ProcessWindowStyle.Minimized;
+                            System.Diagnostics.Process.Start(PSI);
+                        }
                     }
                 }
+                else
+                {
+                    Debug.Log("CashmaticApp", string.Format("File path {0} does not exists", ConfigurationManager.AppSettings["CashmaticAppPath"]));
+                    Global.isCashPayment = false; 
+                }
+               
 
                 this.Content = new TicketScanPage();
